move database code in to a module
