+++ main.py
@@ -6,7 +6,7 @@ from fastapi.staticfiles import StaticFiles
 from rich import print
 
 from config.settings import get_settings
-from db import database
+from database.db import database
 from resources import config_error
 from resources.routes import api_router
